Moving storage Connection behavior into shared class.
